requirements.txt
langchain
ollama
psycopg2-binary
pgvector
google-api-python-client
google-auth-oauthlib
python-dotenv
